+++ controller/users_controller.py
@@ -1,4 +1,6 @@
 from flask import Blueprint, request, jsonify
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
+
 from model.users_model import User
 from schema.users_schema import user_schema, users_schema
 from app import db
@@ -6,14 +8,14 @@ from app import db
 user = Blueprint('user', __name__, url_prefix="/users")
 
 
-# Retrieve all users
+# Retrieve all users (testing only)
 @user.get("/")
 def get_users():
     users = User.query.all()
     return users_schema.dump(users)
 
 
-# get a specific user by id
+# get a specific user by id (testing only)
 @user.route("/<int:user_id>", methods=["GET"])
 def get_user(user_id):
     user = User.query.get(user_id)
@@ -44,11 +46,37 @@ def create_user():
     }
     return jsonify(response), 201
 
+# User login for access token
+@user.route('/login', methods=['POST'])
+def login():
+    if not request.is_json:
+        return jsonify({"msg": "Missing JSON in request"}), 400
+
+    email = request.json.get('email', None)
+    password = request.json.get('password', None)
+
+    if not email:
+        return jsonify({"msg": "Missing email parameter"}), 400
+    if not password:
+        return jsonify({"msg": "Missing password parameter"}), 400
+
+    user = User.query.filter_by(email=email).first()
+    if user and user.password == password:
+        access_token = create_access_token(identity=user.id)
+        return jsonify(message='Login succeeded.', access_token=access_token)
+    else:
+        return jsonify(message='Invalid email or password.'), 401
+
 
 
 # update a specific user by id
 @user.route("/<int:user_id>", methods=["PUT"])
+@jwt_required()
 def update_user(user_id):
+    current_user_id = get_jwt_identity()
+    if user_id != current_user_id:
+        return jsonify({'message': 'Unauthorised request'}), 401
+
     try:
         user = User.query.get_or_404(user_id)
         data = request.json
@@ -62,9 +90,16 @@ def update_user(user_id):
         return jsonify({'message': 'User not found'}), 404
 
 
+from flask_jwt_extended import jwt_required, get_jwt_identity
+
 # delete a specific user by id
 @user.route("/<int:user_id>", methods=["DELETE"])
+@jwt_required()
 def delete_user(user_id):
+    current_user_id = get_jwt_identity()
+    if current_user_id != user_id:
+        return jsonify({'message': 'Not authorised to delete this user'}), 403
+
     try:
         user = User.query.get_or_404(user_id)
         db.session.delete(user)
@@ -72,3 +107,4 @@ def delete_user(user_id):
         return jsonify({'message': 'User deleted successfully'}), 204
     except:
         return jsonify({'message': 'User not found'}), 404
+
